uploader/requirements.txt
pyzmq
structlog
azure-storage-blob
snowflake-connector-python
python-dotenv
